Struct assets::pallet::Pallet

pub struct Pallet<T>(_);

The pallet implementing the on-chain logic.

Implementations

impl<T: Config> Pallet<T>

pub fn register( origin: OriginFor<T>, symbol: AssetSymbol, name: AssetName, initial_supply: Balance, is_mintable: bool, is_indivisible: bool, opt_content_src: Option<ContentSource>, opt_desc: Option<Description> ) -> DispatchResultWithPostInfo

Performs an asset registration.

Registers new AssetId for the given origin. AssetSymbol should represent string with only uppercase latin chars with max length of 7. AssetName should represent string with only uppercase or lowercase latin chars or numbers or spaces, with max length of 33.

pub fn transfer( origin: OriginFor<T>, asset_id: T::AssetId, to: T::AccountId, amount: Balance ) -> DispatchResultWithPostInfo

Performs a checked Asset transfer.

origin: caller Account, from which Asset amount is withdrawn,

asset_id: Id of transferred Asset,

to: Id of Account, to which Asset amount is deposited,

amount: transferred Asset amount.

pub fn mint( origin: OriginFor<T>, asset_id: T::AssetId, to: T::AccountId, amount: Balance ) -> DispatchResultWithPostInfo

Performs a checked Asset mint, can only be done by corresponding asset owner account.

origin: caller Account, which issues Asset minting,

asset_id: Id of minted Asset,

to: Id of Account, to which Asset amount is minted,

amount: minted Asset amount.

pub fn force_mint( origin: OriginFor<T>, asset_id: T::AssetId, to: T::AccountId, amount: Balance ) -> DispatchResultWithPostInfo

Performs an unchecked Asset mint, can only be done by root account.

Should be used as extrinsic call only. Currencies::updated_balance() should be deprecated. Using force_mint allows us to perform extra actions for minting, such as buy-back, extra-minting and etc.

origin: caller Account, which issues Asset minting,

asset_id: Id of minted Asset,

to: Id of Account, to which Asset amount is minted,

amount: minted Asset amount.

pub fn burn( origin: OriginFor<T>, asset_id: T::AssetId, amount: Balance ) -> DispatchResultWithPostInfo

Performs a checked Asset burn, can only be done by corresponding asset owner with own account.

origin: caller Account, from which Asset amount is burned,

asset_id: Id of burned Asset,

amount: burned Asset amount.

pub fn update_balance( origin: OriginFor<T>, who: T::AccountId, currency_id: <<T as Config>::Currency as MultiCurrency<<T as Config>::AccountId>>::CurrencyId, amount: <<T as Config>::Currency as MultiCurrencyExtended<<T as Config>::AccountId>>::Amount ) -> DispatchResult

Add or remove abs(by_amount) from the balance of who under currency_id. If positive by_amount, do add, else do remove.

Basically a wrapper of MultiCurrencyExtended::update_balance for testing purposes.

TODO: move into tests extrinsic collection pallet

pub fn set_non_mintable( origin: OriginFor<T>, asset_id: T::AssetId ) -> DispatchResultWithPostInfo

Set given asset to be non-mintable, i.e. it can no longer be minted, only burned. Operation can not be undone.

origin: caller Account, should correspond to Asset owner

asset_id: Id of burned Asset,

pub fn update_info( origin: OriginFor<T>, asset_id: T::AssetId, new_symbol: Option<AssetSymbol>, new_name: Option<AssetName> ) -> DispatchResultWithPostInfo

Change information about asset. Can only be done by root

origin: caller Account, should be root

asset_id: Id of asset to change,

new_symbol: New asset symbol. If None asset symbol will not change

new_name: New asset name. If None asset name will not change

impl<T: Config> Pallet<T>

pub fn asset_owner<KArg>(k: KArg) -> Option<T::AccountId>where KArg: EncodeLike<T::AssetId>,

Asset Id -> Owner Account Id

impl<T: Config> Pallet<T>

pub fn asset_infos<KArg>( k: KArg ) -> (AssetSymbol, AssetName, BalancePrecision, bool, Option<ContentSource>, Option<Description>)where KArg: EncodeLike<T::AssetId>,

Asset Id -> (Symbol, Name, Precision, Is Mintable, Content Source, Description)

impl<T: Config> Pallet<T>

pub fn asset_infos_v2<KArg>(k: KArg) -> AssetInfowhere KArg: EncodeLike<T::AssetId>,

Asset Id -> AssetInfo

impl<T: Config> Pallet<T>

pub fn tuple_from_asset_id<KArg>(k: KArg) -> Option<AssetRecord<T>>where KArg: EncodeLike<T::AssetId>,

Asset Id -> AssetRecord

impl<T: Config> Pallet<T>

pub fn register_asset_id_from_tuple(tuple: &AssetRecord<T>) -> T::AssetId

Generates an AssetId for the given AssetRecord<T>, and insert record to storage map.

pub fn gen_asset_id_from_any(value: &impl Encode) -> T::AssetId

Generates an AssetId from an Encode value.

pub fn gen_asset_id(account_id: &T::AccountId) -> T::AssetId

Generates an AssetId for the given AccountId.

pub fn register_asset_id( account_id: T::AccountId, asset_id: T::AssetId, symbol: AssetSymbol, name: AssetName, precision: BalancePrecision, initial_supply: Balance, is_mintable: bool, asset_type: AssetType, opt_content_src: Option<ContentSource>, opt_desc: Option<Description> ) -> DispatchResult

Register the given AssetId.

pub fn register_from( account_id: &T::AccountId, symbol: AssetSymbol, name: AssetName, precision: BalancePrecision, initial_supply: Balance, is_mintable: bool, asset_type: AssetType, opt_content_src: Option<ContentSource>, opt_desc: Option<Description> ) -> Result<T::AssetId, DispatchError>

Generates new AssetId and registers it from the account_id.

pub fn ensure_asset_is_mintable(asset_id: &T::AssetId) -> DispatchResult

pub fn transfer_from( asset_id: &T::AssetId, from: &T::AccountId, to: &T::AccountId, amount: Balance ) -> DispatchResult

pub fn mint_to( asset_id: &T::AssetId, issuer: &T::AccountId, to: &T::AccountId, amount: Balance ) -> DispatchResult

pub fn mint_unchecked( asset_id: &T::AssetId, to: &T::AccountId, amount: Balance ) -> DispatchResult

pub fn burn_from( asset_id: &T::AssetId, issuer: &T::AccountId, from: &T::AccountId, amount: Balance ) -> DispatchResult

pub fn update_own_balance( asset_id: &T::AssetId, who: &T::AccountId, by_amount: Amount ) -> DispatchResult

pub fn can_reserve( asset_id: T::AssetId, who: &T::AccountId, amount: Balance ) -> bool

pub fn reserve( asset_id: &T::AssetId, who: &T::AccountId, amount: Balance ) -> DispatchResult

pub fn unreserve( asset_id: &T::AssetId, who: &T::AccountId, amount: Balance ) -> Result<Balance, DispatchError>

pub fn set_non_mintable_from( asset_id: &T::AssetId, who: &T::AccountId ) -> DispatchResult

pub fn list_registered_asset_ids() -> Vec<T::AssetId>

pub fn list_registered_asset_infos( ) -> Vec<(T::AssetId, AssetSymbol, AssetName, BalancePrecision, bool, Option<ContentSource>, Option<Description>)>

pub fn update_asset_type( asset_id: &T::AssetId, asset_type: &AssetType ) -> DispatchResult