Trait assets::pallet::Config

pub trait Config: Config + Config + Config + Config {

    type RuntimeEvent: From<Event<Self>> + IsType<<Self as Config>::RuntimeEvent>;
    type ExtraAccountId: Clone + Copy + Encode + Decode + TypeInfo + Eq + PartialEq + From<Self::AccountId> + Into<Self::AccountId>;
    type ExtraAssetRecordArg: Clone + Copy + Encode + Decode + TypeInfo + Eq + PartialEq + From<AssetIdExtraAssetRecordArg<Self::DEXId, Self::LstId, Self::ExtraAccountId>> + Into<AssetIdExtraAssetRecordArg<Self::DEXId, Self::LstId, Self::ExtraAccountId>>;
    type AssetId: Parameter + Member + Copy + MaybeSerializeDeserialize + Ord + Default + Into<<<Self as Config>::Currency as MultiCurrency<<Self as Config>::AccountId>>::CurrencyId> + Into<AssetIdOf<Self>> + From<AssetIdOf<Self>> + From<AssetId32<PredefinedAssetId>> + From<H256> + Into<H256> + Into<<Self as Config>::CurrencyId> + MaxEncodedLen;
    type GetBaseAssetId: Get<Self::AssetId>;
    type GetBuyBackAssetId: Get<Self::AssetId>;
    type GetBuyBackSupplyAssets: Get<Vec<Self::AssetId>>;
    type GetBuyBackPercentage: Get<u8>;
    type GetBuyBackAccountId: Get<Self::AccountId>;
    type GetBuyBackDexId: Get<Self::DEXId>;
    type BuyBackLiquidityProxy: LiquidityProxyTrait<Self::DEXId, Self::AccountId, Self::AssetId>;
    type Currency: MultiLockableCurrency<Self::AccountId, Moment = Self::BlockNumber, CurrencyId = Self::AssetId, Balance = Balance> + MultiReservableCurrency<Self::AccountId, CurrencyId = Self::AssetId, Balance = Balance> + MultiCurrencyExtended<Self::AccountId, Amount = Amount>;
    type GetTotalBalance: GetTotalBalance<Self>;
    type WeightInfo: WeightInfo;

}

Configuration trait of this pallet.

Implement this type for a runtime in order to customize this pallet.

Required Associated Types

type RuntimeEvent: From<Event<Self>> + IsType<<Self as Config>::RuntimeEvent>

type ExtraAccountId: Clone + Copy + Encode + Decode + TypeInfo + Eq + PartialEq + From<Self::AccountId> + Into<Self::AccountId>

type ExtraAssetRecordArg: Clone + Copy + Encode + Decode + TypeInfo + Eq + PartialEq + From<AssetIdExtraAssetRecordArg<Self::DEXId, Self::LstId, Self::ExtraAccountId>> + Into<AssetIdExtraAssetRecordArg<Self::DEXId, Self::LstId, Self::ExtraAccountId>>

type AssetId: Parameter + Member + Copy + MaybeSerializeDeserialize + Ord + Default + Into<<<Self as Config>::Currency as MultiCurrency<<Self as Config>::AccountId>>::CurrencyId> + Into<AssetIdOf<Self>> + From<AssetIdOf<Self>> + From<AssetId32<PredefinedAssetId>> + From<H256> + Into<H256> + Into<<Self as Config>::CurrencyId> + MaxEncodedLen

DEX assets (currency) identifier.

type GetBaseAssetId: Get<Self::AssetId>

The base asset as the core asset in all trading pairs

type GetBuyBackAssetId: Get<Self::AssetId>

Assets that will be buy-backed and burned for every [GetBuyBackPercentage] of [GetBuyBackSupplyAssets] mints

type GetBuyBackSupplyAssets: Get<Vec<Self::AssetId>>

Assets, [GetBuyBackPercentage] of minted amount of which will be used to buy-back and burn [GetBuyBackAssetId]

type GetBuyBackPercentage: Get<u8>

The percentage of minted [GetBuyBackSupplyAssets] that will be used to buy-back and burn [GetBuyBackAssetId]

type GetBuyBackAccountId: Get<Self::AccountId>

Account which will be used to buy-back and burn [GetBuyBackAssetId]

type GetBuyBackDexId: Get<Self::DEXId>

DEX id to buy-back and burn [GetBuyBackAssetId] through [BuyBackLiquidityProxy]

type BuyBackLiquidityProxy: LiquidityProxyTrait<Self::DEXId, Self::AccountId, Self::AssetId>

Liquidity proxy to perform [GetBuyBackAssetId] buy-back and burn

type Currency: MultiLockableCurrency<Self::AccountId, Moment = Self::BlockNumber, CurrencyId = Self::AssetId, Balance = Balance> + MultiReservableCurrency<Self::AccountId, CurrencyId = Self::AssetId, Balance = Balance> + MultiCurrencyExtended<Self::AccountId, Amount = Amount>

Currency to transfer, reserve/unreserve, lock/unlock assets

type GetTotalBalance: GetTotalBalance<Self>

Get the balance from other components

type WeightInfo: WeightInfo

Weight information for extrinsics in this pallet.